injury_prediction/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.0.33
pyspark==3.3.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2 